requirements.txt
numpy 
pandas
scikit-learn
seaborn
python-dotenv
pymongo
certifi
pymongo[srv]==3.6

##-e .